refactor(probability): isolate Gaussian moment contracts
