+++ src/jacobian/math/probability/_operations.py
@@ -9,6 +9,12 @@ from jacobian._exact import CanonicalRational
 from jacobian.canonical import format_canonical_integer
 from jacobian.catalog._examples import example
 from jacobian.catalog.models import MathTool, MathTools
+from jacobian.math.probability._gaussian import (
+    ExactComplexRational,
+    GaussianMomentContraction,
+    GaussianPolynomialMomentRequest,
+    GaussianPolynomialMomentResult,
+)
 from jacobian.math.probability._gaussian_inputs import (
     CanonicalGaussianPolynomialMomentRequest,
 )
@@ -18,7 +24,6 @@ from jacobian.math.probability._models import (
     DirectedBondConnectionProbabilityResult,
     DirectedBondConnectionProbabilitySource,
     DirectedBondReliabilityState,
-    ExactComplexRational,
     FiniteConditionalContribution,
     FiniteConditionRequest,
     FiniteConditionResult,
@@ -35,9 +40,6 @@ from jacobian.math.probability._models import (
     FiniteRawMomentContribution,
     FiniteRawMomentRequest,
     FiniteRawMomentResult,
-    GaussianMomentContraction,
-    GaussianPolynomialMomentRequest,
-    GaussianPolynomialMomentResult,
     GraphConnectionProbabilityRequest,
     GraphConnectionProbabilityResult,
     GraphReliabilityState,
